Fix warnings from Wall and Wpedantic (#33)

+++ src/GridMaskMax.cc
@@ -73,7 +73,8 @@ void GridMaskMax::applyPrivate(
                 for (int iz = 0; iz < dim2; iz++)
                 {
                     // assert( iy2+iz< gu.grid().size()/subdivx_ );
-                    assert(offset + iy1 + iz < gu.grid().sizeg());
+                    assert(offset + iy1 + iz
+                           < static_cast<int>(gu.grid().sizeg()));
                     ppu[iz] = (T)GridMask::limitAbsValue(ppu[iz], pmask[iz]);
                 }
                 iy1 += incy1;
